[FAB-16217] Do not load JSON Schema schema from network
Currently, the JSON Schema schema (http://json-schema.org/draft-04/schema)
is loaded from the network every time the Java chaincode starts. This has
been seen to cause delays/hangs in the CI system, and we should avoid it
if possible.

This CR adds the JSON Schema schema to the chaincode resources, and uses
that copy (instead of the network copy) to validate any contract schemas.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: Ic559fb5f1a07f86bcd4d2eb2003c791e71abc701