[FAB-16315] Improved Load Ability
A fairly sizeable CR for improving the load handling
ability of the Java chaincode. This required a rework
of the threading model within the core shim layer.

Plus adding configuration and metrics to support this.
The configuration is a simple Java props file should
the need be to ever modify the default thread pool settings.

Metrics are basic stats on the thread pool are written to the
log

Change-Id: I31b05585a0aa650f7e2a7e2b0389799e90adc2c3
Signed-off-by: Matthew B. White <whitemat@uk.ibm.com>
(cherry picked from commit baaaef83d712e3bad460ba34a885169a51ed1474)