[FAB-16655] Warnings as errors for Java chaincode
Enable warnings as errors using -Werror, enable all
linting checks with -Xlint:all,-try, and fix all the
resulting errors.

The -try part is needed to disable an error on the
iterators, which cannot be fixed. It basically says
"you're implementing AutoCloseable, don't you dare
throw InterruptedException!" - which we don't.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: Id1b34b4683fa66039c09905fd3607baeb4292f20