cast type to java.nio.Buffer to flip()
When starting java chaincode as a service with shim 2.2.3 on JDK8, `epochBytes.flip()` throws `java.lang.NoSuchMethodError: java.nio.ByteBuffer.flip()`, which makes the response of the invoke request from peer miss, and peer reports timeout. I try to run the chaincode on JDK9 and it works well. Maybe shim 2.2.3 is built on JDK9+ ?