Improve the handling of the TLS certifiactes
For reasons lost in history, Fabric hasn't standised on whether to load certs in base64 or
PEM format. What does seem to be standard is if the environment variable ends in FILE then
it's a PEM format, but if it's PATH then it's base64.

This change will allow either FILE or PATH environment variables to be used in all cases.

Signed-off-by: Matthew B White <whitemat@uk.ibm.com>