Arbitrary Code Execution Bug
Description:

Affected versions of this package are
vulnerable to Arbitrary Code Execution. A
configuration can be turned on to allow remote
logging through interfaces that accept untrusted
serialized data. Authenticated attackers on the
adjacent network can exploit this vulnerability
to run arbitrary code through the deserialization
of custom gadget chains.

Signed-off-by: Bhaskar <ram@hacker.ind.in>
Signed-off-by: Ry Jones <ry@linux.com>