Use OSV-Scanner instead of dependency-check
The existing dependency-check version is no longer supported and might fail after the NVD data feeds it uses are deprecated on 2023-12-15. The updated version requires an API key to interact with the newer NVD APIs. For details see:

- https://github.com/jeremylong/DependencyCheck#900-upgrade-notice

It also requires periodic triage and suppression of false positive detections. OSV-Scanner appears less prone to false positives and does not require an API key to be maintained.

Implement a scheduled vulnerability scan (using OSV-Scanner) so that vulnerabilities are more visible than the current (dependency-check) implementation, which runs in PR builds but does not fail builds or make the results very visible.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>