Use lighter weight Maven Central publishing plugin (#473)
JReleaser is a large plugin intended to manage the whole project release
lifecycle. It was only being used to publish to the new Maven Central
Publisher API. Some transitive dependencies conflict with those of
Spotless 8.x.

This change switches to the GradleUp nmcp (New Maven Central Publishing)
plugin. It provides only the required publishing capability and is much
lighter weight.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>